burrow package (shared code)
