+++ packages/burrow/tsconfig.json
@@ -7,14 +7,14 @@
     "preserveConstEnums": true,
     "sourceMap": true,
     "strict": true,
-    "target": "ES2020"
+    "target": "ES2020",
+    "composite": true
   },
   "files": [
     "index.ts",
-    "test/index.ts"
-  ],
-  "include": [
-    "commands/**/*.ts"
+    "db.ts",
+    "dice.ts",
+    "utils.ts"
   ],
   "exclude": [
     "node_modules",
