+++ packages/discord-bot/commands/roll.ts
@@ -5,8 +5,10 @@ import { SlashCommandBuilder } from '@discordjs/builders';
 import type { CommandInteraction } from 'discord.js';
 import type { Arena, PlayerId, Reply } from '../api';
 import { lookupArena, lookupPlayerId } from '../api';
-import { redis } from '../db';
-import { Sorry, chooseOne } from '../utils';
+import { redis } from 'burrow/db';
+import { adornName, dailyTrendMarker, increaseDoublerStreak, loadDoubler, multiply, saveNewDoubler } from 'burrow/dice';
+import { Sorry, chooseOne } from 'burrow/utils';
+import { dayRollKey } from 'burrow/dice';
 
 export const data: SlashCommandBuilder = new SlashCommandBuilder()
   .setName('roll')
@@ -216,157 +218,5 @@ function twoSpirit(a: number, b: number, sum: number): string {
       ]);
     default:
       return '';
-  }
-}
-
-export function dayRollKey(arena: string, day: 'today' | 'yesterday'): string {
-  const when = new Date();
-  if (day === 'yesterday') {
-    when.setDate(when.getDate() - 1); // apparently this works
-  }
-  const year = when.getFullYear();
-  const month = when.getMonth() + 1;
-  const date = when.getDate();
-  const leadZero = (n: number) => (n < 10 ? '0' : '');
-  const fullDate = `${year}-${leadZero(month)}${month}-${leadZero(date)}${date}`;
-  return `${arena}:maiden:day:${fullDate}`;
-}
-
-interface AdornParams {
-  name: string;
-  champ: string;
-  brick: string;
-  hundo: string;
-  hundoStreak: number;
-  pooper: string;
-  poopSuite: number;
-  doubler: Doubler;
-}
-
-export const adornName = (params: AdornParams) => {
-  const { name, champ, brick, hundo, hundoStreak, pooper, poopSuite, doubler } = params;
-  const badges = [name];
-  if (!!champ && name === champ) {
-    badges.push('👑');
-  }
-  if (!!hundo && name === hundo) {
-    const token = seasonalToken();
-    badges.push(multiply(token, hundoStreak));
-  }
-  if (!!pooper && name === pooper) {
-    badges.push(multiply('💩', poopSuite));
-  }
-  if (!!doubler.name && name === doubler.name) {
-    badges.push(multiply(doubler.token, doubler.streak));
-  }
-  if (!!brick && name === brick) {
-    badges.push('🧱');
-  }
-
-  return badges.join('');
-};
-
-function multiply(str: string, n: number): string {
-  if (n === 69) return `${str} x69 (nice)`;
-  return n < 20 ? str.repeat(n) : `${str} x${n}`;
-}
-
-const DOUBLES = ['🍒', '✌️', '🫁', '👯', '🤼', '🫂', '🎎', '🙌', '🖇️', '⚔️', '🛠️', '⛓️', '🛍️', '🚻', '👣', '🧦', '🩰', '⚖️', '🧬', '🎵', '♊', '🪺'];
-
-// state for the current doubles-roller
-interface Doubler {
-  name: string;
-  streak: number;
-  token: string;
-}
-
-export async function loadDoubler(arena: string): Promise<Doubler> {
-  const doublerKey = `${arena}:maiden:doubler`;
-  const streakKey = `${doublerKey}_streak`;
-  const tokenKey = `${doublerKey}_token`;
-
-  const name = await redis.get(doublerKey) || '<nobody>';
-  const streak = Number(await redis.get(streakKey));
-  const token = await redis.get(tokenKey) || '✌️';
-  return { name, streak, token };
-}
-
-async function saveNewDoubler(arena: string, name: string): Promise<Doubler> {
-  const doublerKey = `${arena}:maiden:doubler`;
-  const streakKey = `${doublerKey}_streak`;
-  const tokenKey = `${doublerKey}_token`;
-
-  // each doubles streak is assigned a random emoji
-  const token = chooseOne(DOUBLES);
-
-  const tx = redis.multi();
-  tx.set(doublerKey, name);
-  tx.set(streakKey, '1');
-  tx.set(tokenKey, token);
-  await tx.exec();
-
-  return { name, streak: 1, token };
-}
-
-async function increaseDoublerStreak(arena: string, doubler: Doubler) {
-  const doublerKey = `${arena}:maiden:doubler`;
-  const streakKey = `${doublerKey}_streak`;
-  doubler.streak = Number(await redis.incr(streakKey));
-}
-
-function seasonalToken(): string {
-  const today = new Date;
-  const day = today.getDate();
-  const month = today.getMonth() + 1;
-
-  if (month == 1) {
-    if (day <= 7) return '🐣';
-    if (day <= 14) return '🐤';
-    if (day <= 30) return '🐓';
-    return '🍗';
-  } else if (month == 2) {
-    return day <= 14 ? '🍲' : '⛷️';
-  } else if (month == 3 && day <= 23) {
-    return '☕';
-  } else if (month <= 4 && day <= 14) {
-    return '🌸';
-  } else if (month == 4) {
-    // easter weekend 2022
-    if (day <= 17) return '🐇';
-    if (day == 18) return '🍫';
-    if (day <= 25) return '🌱';
-    return '🪴';
-  } else if (month == 5) {
-    if (day <= 7) return '🌳';
-    if (day <= 14) return '🐛';
-    if (day <= 21) return '🦋';
-    return '🦆';
-  } else if (month == 6) {
-    if (day <= 7) return '🌊';
-    if (day <= 14) return '🏄';
-    if (day <= 21) return '🏝️';
-    return '🪸';
-  } else if (month == 7) {
-    if (day == 1) return '🍁';
-    if (day <= 7) return '🐚';
-    if (day <= 14) return '🦐';
-    if (day <= 21) return '🦩';
-    return '🌻';
-  } else if (month == 8) {
-    return '🦗';
-  }
-  // TBD...
-  if (month == 12 && day > 25) {
-    return '🥚';
-  }
-  return '⛺';
-}
-
-function dailyTrendMarker(daily: 'new day' | 'higher' | 'lower' | undefined): string {
-  switch (daily) {
-    case 'new day': return ' ☀️';
-    case 'higher': return ' 📈';
-    case 'lower': return ' 📉';
-    default: return '';
   }
 }
