+++ packages/web/package.json
@@ -11,6 +11,7 @@
     "test": "tsc --noEmit"
   },
   "dependencies": {
+    "burrow": "workspace:*",
     "next": "12.1.6",
     "react": "18.1.0",
     "react-dom": "18.1.0"
